feat: add url with custom page transition to history (#489)
